Split color_loader module
Move method definitions to a module to minimise inclusion overhead.
