+++ src/color_loader.h
@@ -32,55 +32,25 @@ class color_loader
             return names;
         }
 
-        void load_colors( const JsonObject &jsobj ) {
-            for( size_t c = 0; c < main_color_names().size(); c++ ) {
-                const std::string &color = main_color_names()[c];
-                JsonArray jsarr = jsobj.get_array( color );
-                consolecolors[color] = from_rgb( jsarr.get_int( 0 ), jsarr.get_int( 1 ), jsarr.get_int( 2 ) );
-            }
-        }
-        ColorType ccolor( const std::string &color ) const {
-            const auto it = consolecolors.find( color );
-            if( it == consolecolors.end() ) {
-                throw std::runtime_error( std::string( "requested non-existing color " ) + color );
-            }
-            return it->second;
-        }
-
-        void load_colorfile( const std::string &path ) {
-            std::ifstream colorfile( path.c_str(), std::ifstream::in | std::ifstream::binary );
-            JsonIn jsin( colorfile );
-            jsin.start_array();
-            while( !jsin.end_array() ) {
-                JsonObject jo = jsin.get_object();
-                load_colors( jo );
-                jo.finish();
-            }
-        }
+        void load_colors( const JsonObject &jsobj );
+        ColorType ccolor( const std::string &color ) const;
+        void load_colorfile( const std::string &path );
 
     public:
         /// @throws std::exception upon any kind of error.
-        void load( std::array<ColorType, COLOR_NAMES_COUNT> &windowsPalette ) {
-            const std::string default_path = PATH_INFO::colors();
-            const std::string custom_path = PATH_INFO::base_colors();
-
-            if( !file_exist( custom_path ) ) {
-                copy_file( default_path, custom_path );
-            }
-
-            try {
-                load_colorfile( custom_path );
-            } catch( const JsonError &err ) {
-                debugmsg( "Failed to load color data from \"%s\": %s", custom_path, err.what() );
-
-                // this should succeed, otherwise the installation is botched
-                load_colorfile( default_path );
-            }
-
-            for( size_t c = 0; c < main_color_names().size(); c++ ) {
-                windowsPalette[c] = ccolor( main_color_names()[c] );
-            }
-        }
+        void load( std::array<ColorType, COLOR_NAMES_COUNT> &windowsPalette );
 };
+
+class SDL_Color;
+class RGBQUAD;
+
+template<>
+SDL_Color color_loader<SDL_Color>::from_rgb( const int r, const int g, const int b );
+template<>
+RGBQUAD color_loader<RGBQUAD>::from_rgb( const int r, const int g, const int b );
+
+extern template class color_loader<SDL_Color>;
+extern template class color_loader<RGBQUAD>;
+
 
 #endif // CATA_SRC_COLOR_LOADER_H
